backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-colorfield==0.9.0 
django-cors-headers==3.13.0
django-filter==23.2
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
